Add update and existance checking to Lists.js

+++ src/services/tables/Lists.js
@@ -4,7 +4,7 @@ export default class Lists {
     }
 
     async all() {
-        return await this.conn.execute(
+        return await this.conn.select(
             "SELECT * FROM lists"
         );
     }
@@ -16,29 +16,33 @@ export default class Lists {
         );
     }
 
-    async getOne(id) {
-        return await this.conn.select(
-            "SELECT * FROM lists WHERE id=$1",
-            [id]
-        );
-    }
-
-    async getBoard(board_id) { // TODO: move to Db function as join?
-        return await this.conn.select(
-            "SELECT * FROM lists WHERE board_id=$1",
-            [board_id]
-        );
+    async update(id, board_id, title, color) {
+        // Note, not to be used to update position
+        return await this.conn.execute(
+            "UPDATE lists SET board_id=$1, title=$2, color=$3 WHERE id=$4",
+            [board_id, title, color, id]
+        )
     }
 
     async remove(id) {
-        const toRemove = await this.getOne(id);
+        const toRemove = await this.conn.select(
+            "SELECT position FROM lists WHERE id=$1",
+            [id]
+        );
+        if (toRemove.length == 0) {
+            throw new Error(`Removing list ${id}, which does not exist`);
+        }
+        if (toRemove.length > 1) {
+            throw new Error(`Removing list ${id}, which matches multiple rows`);
+        }
+        const positionRemoving = toRemove[0].position
         const res = await this.conn.execute(
             "DELETE FROM lists WHERE id=$1",
             [id]
         );
         await this.conn.execute(
             "UPDATE lists SET position = position - 1 WHERE position > $1",
-            [toRemove.position]
+            [positionRemoving]
         );
         return res;
     }
